+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to the "quokka-alias" extension will be documented in this file.
 
+## [0.2.0] - 06-09-2023
+
+- Now uses the nearest `tsconfig.json` and `.quokka` to resolve aliases
+- Now traverses the entire import tree to resolve aliases
+
 ## [0.1.0] - 2023-04-25
 
 - Initial release